+++ nctest/vputget.c
@@ -77,8 +77,8 @@ test_varputget(cdfid)
 	for (ie = 0; ie < ne; ie++) { /* for each of ne points */
 
 	    /* allocate space for the cube of values */
-	    hc[ie].vals = emalloc(nel[ie]*nctypelen(test.vars[iv].type) + 8);
-	    tmp.vals = emalloc(nel[ie]*nctypelen(test.vars[iv].type) + 8);
+	    hc[ie].vals = emalloc((size_t)(nel[ie]*nctypelen(test.vars[iv].type) + 8));
+	    tmp.vals = emalloc((size_t)(nel[ie]*nctypelen(test.vars[iv].type) + 8));
 
 	    /* fill allocated space with different values of right type */
 	    val_fill(test.vars[iv].type, nel[ie], hc[ie].vals);
